tests/yggdrasil/requirements.txt
pytest>=8.0.0
requests>=2.31.0
Pillow>=10.0.0
cryptography>=42.0.0
